Lives are also saved in the cookie.
This is to prevent cheating. If the user loses a life, he could just refresh
then load game and play the level with 3 lives. But now the lives left are stored,
although the game map is not, so he has to play the entire level again.
XXX: is this a penalty for reloading or should we save the game map?